breaks out single layer component
