+++ client/components/index.js
@@ -1,2 +1,3 @@
 export { default as App } from './App';
 export { default as ConnectionInfo } from './ConnectionInfo';
+export { default as SingleLayer } from './SingleLayer';
